Check for __amd64__, not __x86_64__.
Spiritual revert of 1fa4de80fcfc697b5e5879cc351fb3e9dbf6acbe.  Intel's C
compiler claims to be gcc-compatible; if they're not defining the same
macros as gcc then that's their bug, not ours.  Even if we were to do
this aliasing we should do it once and for all in servermd.h.