configure.ac: Fix SHA1 handling
Fix and marginally simplify the SHA1 handling.  First, we allow people
to override it.  Secondly, we try for libmd.  Then, we try for OpenSSL
with pkg-config.  In a last, desperate move, we try libcrypto on its
own.  This allows the server to, y'know, _link_ when using OpenSSL,
instead of failing because we only have -lcrypto, and not -lssl.