fbdevhw: Remove pixclock check.
The check can fail because the output from FBIOGET_VSCREENINFO is used to set
Clock in fbdev2xfree_timing().  Then in fbdevHWSetMode(), xfree2fbdev_timing()
is called which sets the pixclock based on Clock.  The resulting circle results
in slight rounding errors, causing the comparision check in fbdev_modes_equal
to fail.