mi: always update the sprite for master devices.
Follow-up to 9ce995373e4a. This re-enables cursor rendering if the MD is
controlled through software (e.g. synergy).

Reported by John Tapsell: "I use Xorg with no mouse attached, but use synergy
to control the mouse.  The commit means that I no longer have a visible mouse
cursor.  The mouse cursor is still 'there' in terms that I can click buttons
etc with it, but it's just not visible."

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>