XKB: Add XkbRMLVOSet
XkbRMLVOSet is just a set of strings for rules, model, layout, variant
and options; use that in preference to XkbRF_VarDefsRec, which is a
hideously complicated monster that somehow managed to not include the
actual rules.

While we're at it, clean up xkbrules.h so it doesn't require xkbstr.h.

Signed-off-by: Daniel Stone <daniel@fooishbar.org>