Remove extra / from Loading module line
In the X log, upon module load, it prints a line similar to the following.

> (II) Loading /usr/lib64/xorg/modules/extensions//libdbe.so

The attached patch removes the extra / before the module name.

Code already exists in hw/xfree86/loader/loadmod.c:InitPathList to add a
trailing slash if needed, removing the one added by sprintf is harmless.

Signed-off-by: James Cloos <cloos@jhcloos.com>