XQuartz: xpr: The dri.c code for pixmaps was wrong in several ways.  They weren't being exported correctly by Xplugin.
This should fix a bug with the surface for a window, when an export fails.
Before the export could fail and leave behind an invalid (freed) pointer in the dix privates.

I have an idea of how to fix the GLXPixmaps now without using CGLSetOffScreen.

This work is a step towards that.  The Xplugin will need a small patch to fix an
issue that this change brought forth.
(cherry picked from commit 58c4116c47543b5e30c2232e7bee8efc0b9be176)