Oops. 62fc98c had crtc offsets applied backwards (fix from server-1.6-branch)
Crtc offsets are in screen space, not crtc space, and hence should be
applied relative to the screen->crtc transform, not the crtc->transform.

This fix was found in the 'cherry pick' of 62fc98c on server-1.6-branch,
clearly some new definition of 'cherry pick' that I am unaware of.

Signed-off-by: Keith Packard <keithp@keithp.com>